% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block}
\alias{block}
\title{block}
\usage{
block(namefield, namespace, pathRes)
}
\arguments{
\item{namefield}{name of the field}

\item{namespace}{namespace of the package}

\item{pathRes}{path of the results}
}
\description{
block
}
\keyword{onionR}

